requirements/tests.txt
-r common.txt
coverage==4.0.3
model-mommy==1.2.6
psycopg2==2.6.1
py==1.4.31
pytest==2.8.7
pytest-cov==2.2.0
pytest-django==2.9.1
pytz==2015.7
